Move tutorial and simplify dash
